requirements.txt
jira==3.4.1
apache-airflow-providers-mongo==3.0.0